+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## [Unreleased]
 
-* Added `save_channels` option.
+* Added `save_channels`, `save_channels_format` and `save_channels_style` options.
 
 * Bug fixes